Remove maximumchars rule for 'alternative_hosts' activity instance setting
This patch removes a maximumchars rule for the 'alternative_hosts' field within the activity instance settings due to complete uselessness.

This rule was introduced with commit 58e27d8, but it was broken from the very beginning as it added a rule to the 'name' field instead of the 'alternative_hosts' field.
Additionally, when the length restriction of the 'alternative_hosts' field was removed in commit 350d0d0, this rule was not removed as well but remained in the activity instance settings.